Component Cleanup - Move Login/Register into auth folder - Remove duplicate login/register from before - Fix imports
