+++ client/src/App.js
@@ -2,9 +2,8 @@ import './App.css';
 import React, { Fragment, useEffect } from 'react';
 import Landing from './components/layout/Landing'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import Login from './components/auth/Login'
-import Login2 from './components/login/Login_Container'
-import Register from './components/auth/Register'
+import Login from './components/login/Login_Container'
+import Register from './components/auth/login/Register'
 import Alert from './components/layout/Alert'
 //Redux
 import { Provider } from 'react-redux';
@@ -41,7 +40,6 @@ const App = () => {
           <Switch>
             <Route exact path='/register' component={Register} />
             <Route exact path='/login' component={Login} />
-            <Route exact path='/login2' component={Login2} />
           </Switch>
         </Fragment>
       </Router>
