--- client/src/components/auth/Login.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Login = () => {
-    return (
-        <div>
-            Login
-        </div>
-    )
-}
-
-export default Login
