--- client/src/components/auth/Register.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Register = () => {
-    return (
-        <div>
-            Register
-        </div>
-    )
-}
-
-export default Register
