+++ client/src/components/auth/login/Register.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
 import { connect } from 'react-redux'
-import { setAlert } from '../../actions/alert'
-import { register } from '../../actions/auth'
-import NbaImage from '../../Assets/Svg/nba.svg'
+import { setAlert } from '../../../actions/alert'
+import { register } from '../../../actions/auth'
+import NbaImage from '../../../Assets/Svg/nba.svg'
 import './style.scss'
 import PropTypes from 'prop-types'
 import { Redirect } from 'react-router-dom'
